Refactor modules to introduce library crate
